fruitAGNTCY/fruit_agents/fruit_cognition/pyproject.toml
[project]
name = "fruit_cognition"
version = "0.1.0"
description = ""
authors = [{ name = "Shridhar Shah", email = "shridhsh@cisco.com" }]
requires-python = ">=3.13,<3.14"
dependencies = [
    "a2a-sdk==0.3.20",
    "click>=8.1.8",
    "coloredlogs>=15.0.1",
    "dotenv>=0.9.9",
    "fastapi>=0.116.0",
    "httpx>=0.23.0",
    "langchain-anthropic>=0.3.13",
    "langchain-google-genai>=2.1.4",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.1",
    "langgraph-supervisor>=0.0.26",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "requests",
    "starlette>=0.49.1",
    "uvicorn>=0.29.0",
    "mcp[cli]>=1.10.0",
    "ioa-observe-sdk==1.0.34",
    "agntcy-identity-service-sdk==0.0.7",
    "llama-index-llms-azure-openai==0.4.2",
    "langchain>=0.3.27",
    "langchain-litellm==0.3.5",
    "llama-index-llms-litellm==0.6.3",
    "llama-index-core==0.14.9",
    # litellm 1.83.0 fixes SQL injection in proxy API key verification (CVE applies to library users).
    # Capped <1.83.6 because 1.83.6+ pins jsonschema==4.23.0, conflicting with openapi-spec-validator (dev).
    # The remaining 1.83.7+ fixes are litellm-proxy server vulnerabilities; fruit_cognition uses litellm as a library, not as a proxy.
    "litellm>=1.83.0,<1.83.6",
    "google-adk>=1.28.1",
    "slima2a==0.3.0",
    "slim-bindings==1.1.0",
    "agntcy-app-sdk==0.5.4",
    "azure-core==1.38.0",
    # Vulnerability fixes
    "langchain-core>=1.2.28",
    "urllib3>=2.6.3",
    "aiohttp>=3.13.4",
    "pyasn1>=0.6.3",
    "pypdf>=6.10.2",
    "filelock>=3.20.3",
    "marshmallow>=3.26.2",
    "PyNaCl>=1.6.2",
    "mcp>=1.23.0",
    "cryptography>=46.0.7",
    "jsonschema>=4.0.0",
    "referencing>=0.28.0",
    "nltk>=3.9.4",
    "authlib>=1.6.11",
    "pillow>=12.2.0",
    "PyJWT>=2.12.0",
    # pyopenssl: cannot bump to 26+ — agntcy-dir==1.0.0 (dev) transitively pins pyopenssl<26 via spiffe-tls.
    # The DTLS cookie callback buffer overflow does not affect fruit_cognition (no DTLS usage).
    "python-multipart>=0.0.26",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.0.10",
    "langsmith>=0.7.31",
    "langchain-openai>=1.1.14",
    "Mako>=1.3.11",
    "Pygments>=2.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=0.23.0,<0.24",
    "pytest-cov>=4.0.0,<5",
    "typing-extensions>=4.12.2,<5",
    "openai>=2.8.0,<3.0",
    "coloredlogs>=15.0.1,<16",
    "langchain-openai>=1.1.14",
    "sentence-transformers>=5.1.1",
    "agntcy-dir==1.0.0",
    "openapi-spec-validator>=0.8.4",
    "prance>=25.4.8.0",
    "pyyaml>=6.0.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[[tool.uv.index]]
url = "https://buf.build/gen/python"

[tool.poetry]
package-mode = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["farm", "exchange", "schema", "api"]

[tool.pytest.ini_options]
addopts = "--asyncio-mode=auto"
filterwarnings = [
    "error",
]
markers = [
    "agents(names): marks tests that run for specific farms (e.g. @pytest.mark.farms([\"brazil\"]))",
]
pythonpath = ["."]
testpaths = ["tests"]
